% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aprox.R
\name{aprox}
\alias{aprox}
\title{Numerical approximation of defined integrals.}
\usage{
aprox(func, n_points, interval, mode = "s")
}
\arguments{
\item{func}{The function that is gonna be integrated. The function needs to
return a single number as input and receive a single number as input.}

\item{n_points}{An integer for the number of points that are gonna be used
for the  approximation. More points will generate a better approximation but
will be heavier computationally}

\item{interval}{A length 2 numeric vector that is the interval to integrate}

\item{mode}{The mode used for approximation, the available modes are:
- "m" : Midpoint method
- "l" : Left endpoint method
- "r" : Right endpoint method
- "t" : Trapezoidal method
- "s" : Simpsons method}
}
\value{
A Numeric. The area bellow the function func in the interval provided.
}
\description{
A function for calculating approximations of defined integrals with various
different methods. The methods available are explained in the details section.
}
\details{
The integration methods available are:
  - "m": Midpoint method
  - "r": Right endpoint method
  - "l": Left endpoint method
  - "t": Trapezoidal method
  - "s": Simpson's rule
}
\examples{
aprox(function(x) x^2, n_points = 100, interval = c(0,1))
aprox(function(x) sqrt(x^x + 2), n_points = 100, interval = c(sqrt(2), pi), mode = "m")
aprox(function(x) exp(x^2), n_points = 1000, interval = c(2,4), mode = "t" )
aprox(function(x) exp(x^2), n_points = 1000, interval = c(4,2))
}
